requirements.txt
flask[async]==2.0.1
werkzeug<2.3.0
requests==2.28.1
youtube-transcript-api==0.5.0
pytube==12.1.0
SpeechRecognition==3.9.0
pydub==0.25.1
openai-whisper>=20230918
torch>=2.0.0
langdetect==1.0.9
mcp[cli]>=0.7.0
asgiref>=3.2.10